requirements.txt
fastapi
uvicorn
python-dotenv
pipecat-ai[webrtc,daily,deepgram,cartesia,silero]
pipecat-ai-small-webrtc-prebuilt
loguru